ubertest: Require use of -u option
Ubertest requires that either a config file be provided as input
(-u option) or that the provider and test name be given (-p and
-t options).  This is confusing, particularly since there's no
information given as to what the test name could be or how it is
used to locate a file.  Remove the second option and just require
that the config file be provided.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>