rbtree: Free the specified node from the rbtree
The rbtree delete call takes as input a node to remove from the
tree.  As an optimization, that node may be left in place in the
tree, with only the data it references replaced.  Instead, another
node in the tree (the node that is logically immediately to the
right of the given node) may be delete.  But the data values
of the two nodes are swapped.

This can lead to a possible use after free issue.

The issue results from the rbtree insert call returning a
pointer to the node where new data was inserted.  If that node
pointer is saved but later accessed, it could have been freed.
The freeing of that node is separate from the data originally
associated with the node being removed from the tree because
of the above optimization.

The typically 'fix' for this is that an rbtree delete call must
be preceeded by a find call to locate the node currently
associated with the data/key to remove.  This works fine for
cases where simple key values are used, such as an integer or
an address, where comparison operators use >, <, memcmp, etc.

However, for more complex keys, such as those used by the MR
cache, this presents an issue.  The MR cache uses a comparison
function that involves comparing addresses and length fields
from an iovec.  And to make a long story short, this can result
in the find call matching different nodes, depending on the
state of the tree, even when no entries have been removed
from it.

Because the node->data stored by the rbtree corresponds to a
specific mr_entry allocated by the cache, when we free an
mr_entry, we need to match the specific node where the
node->data == mr_entry.  However, because of the complexity
of the comparison function, that is no longer guaranteed.
This can show up when: merging regions is disabled, we have
overlapping regions in the cache, and we have a third region
that falls within both overlapping regions.  In this case, the
third region can match either region, depending on the order
of their insertion.

To avoid this issue, we need to store the node that was
allocated for a new key inserted into the rbtree.  That exact
same node must be deleted when removing the key.  This allows
us to have a 1:1 relationship between an mr_entry and a node
in the tree.

The insert call already returns the node to the caller.  However
we need to update the delete call to free that node, because
we can't rely on calling find prior to deleting the key.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>