core/rbtree: Modify ofi_rbmap_insert to return the new node
When a new item is inserted into the rbtree, it is associated
with an ofi_rbnode.  That node uniquely identifies the inserted
item.  If requested, return the node to the caller on a
successful insertion.  This will be needed in some use cases
to remove the item from the tree later.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>