man/fence: Expand completion semantics of FI_FENCE ops
FI_FENCE is defined such as previous operations must have
completed prior to the fenced operation.  However, no
requirement is defined for the strength of those completions.
For example, if an operation requires FI_INJECT_COMPLETE, but
the fenced operation needs FI_DELIVERY_COMPLETE, the state
of the first operation is not known.

Expand the fence definition such that all previous operations
must meet the same completion semantic as the fenced
operation.  This is basically the only definition that makes
sense for fencing to be usable.  Just specify this in the
man pages.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>