fabric/collectives: Assign an address to an AV set
fi_join_collective() performs a collective operation itself with
the endpoints identified via the coll_addr parameter.  This
works well for all join operations except the first.  In the case
of the first join, the user passes in ADDR_UNSPEC.  This indicates
that all endpoints that are part of the associated AV are
participarint in the collective.

From the viewpoint of the app, this is restrictive and a little
unusual.

Instead of creating a special case to handle the first join
call, associate an fi_addr with an AV set.  Applications can
create an AV set, populate it with peer addresses, then obtain
an fi_addr for that set of peers.  The fi_addr returned by the
AV set is a local operation only.

By creating an AV set that contains all addresses in an AV,
we maintain the existing support.  But, we also provide a
optimal mechanism for creating collective groups without
needing all endpoints in an AV to participate in the join.
(MPI may not take advantage of this, but other apps might).

Update man pages with new definition, plus include other
missing documentation.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>