collectives: Rework collective APIs to add gather / reduce support
Update the collective APIs to support gather and reduce
operations.  These were missing from the initial collective
calls.  Adding these exposed an issue with the existing calls
providing the 'root' of broadcast calls, which are also
needed for gather and reduce.

To support the new calls, and make it easier to extend support
for other collective ops, modify the calls into the provider.
Instead of using a single writeread() atomic style call, define
separate entry points for each collective.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>