util/uffd: Unregister the reported region for REMOVE events
According to the kernel documentation for userfaultfd
(admin-guide/mm/userfault.rst), when an UNMAP event is
reported, the userfaultfd context "disappears".  We
intepret this to mean that there is no need to unregister
the region associated with the UNMAP event, since the
address range is no longer valid.

However, the same is not true for REMOVE events.  To
simplify the behavior of the MR cache, mask the difference
between REMOVE and UNMAP events.  When a REMOVE event
is reported, unregister it.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>