mr/cache: Move unsubscribe to align with caching
When a new memory region is added, we subscribe for events
only if the entry is being cached.  And if subscription
fails, we uncache the entry.  However, we keep the
subscription until the entry is freed.

To better align the behavior, move unsubscribe to when
the entry is removed from the cache.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>