ubertest: Only test with the first returned fi_info
If fi_getinfo returns multiple fi_info structures, the
server iterates over all returned fi_info's.  As a result
the number of tests that are actually run is multiplied
by the number of fi_info's.  Since there could be a
different fi_info for each address and/or interface, the
number of tests being run can increase by a factor of 4
or more.

This level of testing is unneeded and not the best way to
handle multiple interfaces/addresses being available.
Instead, addition input should be used to select which
interface to test, such as the domain name or an fi_info
index.

Because of the structure of the changes, we make a simple
update to only keep the first fi_info structure and free
the rest.  This ensures that we only run one test iteration
regardless of how many interfaces are available.

This is not an ideal solution.  An ideal patch would remove
the loop iterations in ft_fw_process_list_server.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>