ubertest/psm2: Remove SENDV bandwidth testing
Ubertest re-uses the same data buffer for all sends and receives.
This exposed an issue in the psm2 provider.  The provider posts
the start of the data buffer in order to receive the underlying
protocol header.  After parsing the header, it posts the buffer
to receive the rest of the data.

With bandwidth tests, a stream of messages are received.  When
a single buffer is posted to receive the messages, the received
header is overwritten.  This results in the test hanging.

Since it's highly unlikely that a real application is going to
reuse the same buffer to receive all messages, the simplest
solution is to remove this specific test for the psm2
provider.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>