prov/util: Remove monitor unsubscribe from cache notify
See Issue #5347 for more details. Due to userfault unsubscribing from
monitored regions automatically in the notify path, calling unsubscribe
could unregister regions that still need to be monitored.

There are some other possible race conditions that can reproduce this
issue. Until a subscription context is added, we must remove the
unsubscribe call. This will result in extra notifications, but is
harmless to operation.

Signed-off-by: William Zhang <wilzhang@amazon.com>