prov/util: Refactor queuing of collective work items
The initial implementation used 3 queues to track work items
  1. deferred queue - tracks work items for a collective
  2. pending queue - tracks work items waiting to transfer remotely
  3. barrier queue - tracks work items which block collective progress

This commit reduces to 2 queues and improves fairness in how the work items
among multiple collectives are processed, while removing some lock latency.
  1. util_ep.coll_ready_queue - tracks work items which are ready for processing
  2. util_coll_operation.work_queue - tracks waiting work items for a collective

As work items from the coll_ready_queue complete, the ep will initiate the next
waiting work item for the collective by adding a work_queue item to the
coll_ready_queue. If a ready-to-work item requires remote transfer and is blocked
by the ep for any reason, the item goes to the end of the ready queue and will be
retriggered rather than attempting the same item repeatedly.  This ensures that
multiple collectives get a chance to progress.

Signed-off-by: Stephen Oost <stephen.oost@intel.com>