Fix gcc-9.2.1 stringop warnings
Fixes issue #5386. gcc 9 adds compiler warnings for strncpy
overflows and NUL termination. This fixes the two warnings
that occurred when compiling the existing code. The change
in util_shm.c just suppresses the warning; the existing code
was correct. The change in sock_av.c does fix a possible
overflow in the non-debug case.

Signed-off-by: John Byrne <john.l.byrne@hpe.com>