prov/efa: prov/efa: Initialize global objects for DSO builds
76a9f40ffe added logic to mark pinned memory with MADV_DONTFORK to
prevent the copy-on-write semantics in the face of a fork. Given the
pinned memory could be backed by regular or huge pages, efa_madvise()
relied on the global page_sizes array to try marking the region with all
known sizes. When the provider is built as a DSO library, the global
page_sizes array is uninitialized causing the function to fail and
return an erroneous errno. This commit fixes it by calling
`ofi_mem_ini()` explicitly when initializing the EFA provider in a DSO
build.

Likewise, the MR cache code in EFA relies on the global monitor which
needs to be initialized inside a DSO build.

Signed-off-by: Raghu Raja <craghun@amazon.com>