prov/shm: add user ID to shared memory name
The current implementation does not account for different users using
the shm provider on the same system with the same default settings.

As a result, the second user will fail smr_create because the name
is already in use. Or, if the first user's application resulted in
and error and the shared memory region was not cleaned up properly,
the second user will not be able to remove it and run their its own
application.

To fix this, add the user ID to the beginning of the default endpoint
name to allow different users to run their own applications
independently.

Signed-off-by: aingerson <alexia.ingerson@intel.com>