prov/shm: Cleanup tmpfs cruft after an unclean shutdown
Maintain a global list of smr EP names and use that to unlink the stray
files from the signal handlers registered on init. Currently, this is
the only action we take from the handler for each of the registered
signals. If we run into applications that are smart enough to recover
from a SIGBUS in their own signal handler, there is one corner-case
sequence that might make their recovery moot:

- SIBUG handled in shm and file unlinked
- App recovers from SIGBUS in the old_action handler
- If all peers discovered each other prior to the SIGBUS and mapped in
  the shm file, life is good
- If a peer tries to map in a unlinked file after a recovery gets an
  ENOENT.

The window of this happening is too small. If we do see an app that hits
this, we can avoid catching SIGBUS altogether.

Signed-off-by: Raghu Raja <craghun@amazon.com>