util/shm: round tx/rx sizes up
These sizes are used for both the shared memory region
as well as the cmd queues. The utility queue code rounds
the sizes up, but the shared memory region size calculation
does not causing an issue if the sizes passed in are not
a power of 2. Round this when it is first passed to guarantee
any calls after this that might round the sizesare already
are using a power of 2

Signed-off-by: aingerson <alexia.ingerson@intel.com>