prov/rxm: Remove requirement forcing core to provide mr key
When opening a domain, the rxm provider modifies the mr_mode
in order to force the core provider to set the mr key.  This
has the effect that the rxm provider may be set to MR_SCALABLE,
but the core provider will be set to MR_BASIC.

However, RMA ops pass through to the core provider, which
means that rxm will use 0-based offsets (exposed to the app),
but the core would use virtual address based offsets.

This fixes hangs in shmem testing with rxm + tcp.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>