prov/rxm: Specify mr key when registering memory
The provider currently requires that the underlying provider
set the mr key when registering buffers.  However, this
breaks apps when rxm and the core providers mismatch on
using MR_BASIC vs MR_SCALABLE.  As a result, a follow on
patch will fix this.  However, in order to apply that patch,
we need to prepare rxm to provide the mr key.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>