prov/verbs: Only enable SQ overflow optimization on mlx
We currently disable a SQ/CQ overflow optimization for
qib and hfi1 devices.  However, there's no guarantee that
other devices implement the same behavior that allows
disabling the credit checks.  At this point, only mlx
devices (vendor id 0x02c9) support this.  Reverse the
check, so that the optimization is not the default, but
credit tracking is.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>