prov/rxm: Fix inject call completion semantics
The inject calls allow for immediate reuse of the user's buffer.
It does not change the behavior of when the operation completes.
If the inject operation is associated with a completion counter,
for example, and the endpoint is configured with
FI_DELIVERY_COMPLETE, the counter should not increment until
the transfer has been processed by the peer.

Map inject calls associated with counters so that we generate a
completion.  Update any counters from the inject call as part of
completion handling (code already exists for this).

This fixes an issue running shmem collective tests over the
rxm+tcp providers, where data corruption would be detected
(because of slowness updating the data).

Signed-off-by: Sean Hefty <sean.hefty@intel.com>