prov/efa: use FI_UNIVERSE_SIZE for AV sizing
Read FI_UNIVERSE_SIZE during the AV open. We need to do this as we enforce a
minimum AV size and will skip the universe size read during the util AV open.

We also need to size the reorder buffer freestack so use util AV size so that
FI_UNIVERSE_SIZE is implicitly used. A follow up patch will change this to a
bufpool instead of a freestack so that it can grow indefinitely.

Signed-off-by: Robert Wespetal <wesper@amazon.com>