prov/verbs: fix regression exposed in d38c886
Fix the initialization of fi_nic vendor/device id string, the
current initialization was inconsistent with norms for PCI ID's
and other provider initializations; causing the check in
the commit creating the regression to fail.

Add an assert when specifying tracking post/poll functions
to catch that an invalid code path has been taken.

Signed-off-by: Steve Welch <swelch@systemfabricworks.com>