prov/usnic: always return a domain name
In pre-1.4, the usnic provider domain name was NULL.

There used to be elaborate schemes to try to preserve this pre-1.4
behavior (i.e., only provide a domain name in >=v1.4).  In Nov 2019
discussions, however, it was determined that we could rationalize
classifying this pre-1.4 behavior as buggy.  Specifically: we should
just now always return a domain name -- even if the requested version
is <1.4.

This greatly simplifies the logic here, and also greatly simplifies
layering with the rxd provider.

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>