prov/efa: handle msg_id wraparound correctly
The msg_id's between the sender and receiver were different ranges as
the sender skipped UINT_MAX but the reorder buffer code expected
UINT_MAX. Change the range to be 0 to UINT_MAX. We'll no longer be able
to use UINT_MAX as an invalid msg_id during tx_entry init but the
tradeoff here is that we will overcomplicate the reorder buffer code
trying to special case this.

In addition, the sender was using msg_id 0 twice after wrapping around
which is fixed here.

Finally, update the error print and check for invalid message IDs.

Signed-off-by: Robert Wespetal <wesper@amazon.com>