prov/rxm: Use flag to indicate state of progress thread
Add a flag to the rxm_ep that will be read by the progress
thread.  By clearing the flag, the progress thread will
know to exit, rather than sitting in a while (1) loop.
This avoids the progress thread running forever in the case
that some other thread reads the 'exit' signal event.

As part of the change, move the EXIT handling into the
main FI_NOTIFY handling.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>