prov/verbs: Move control loop into fi_ibv_poll_outstanding_cq
The purpose of this call is to flush all completions from the
underlying verbs CQ.  As indicated by a comment in the function,
it is aware that it is being called in a loop.  Move that
loop into the function and rename it to reflect what it is
trying to accomplish.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>