prov/verbs: Move discarding of internal completions into poll_cq
There are only a couple of places where vrb_poll_cq is now
called, and they all check for internal completions.  We
can simplify the code by moving that check into vrb_poll_cq
and looping until either the CQ is empty or we retrieve
a user completion.

This eliminates the need for vrb_get_user_comp.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>