prov/verbs: Serialize CQ credits with lock rather than atomic
The cq credit field is defined as an atomic.  Instead of using
an atomic value, serialize access to the value using the CQ
lock.  The CQ lock may be a no-op in certain cases, and we will
need the lock to serialize access to a buffer pool in a
follow on patch, making the atomic unnecessary.

Note that the comment above CQ credits is incorrect.  When
multiple endpoints are bound to the same CQ, we can overrun
the CQ without overruning an individual send queue.  Just
remove the comment, as we will always need a credit when
sending.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>