prov/shm: include string NULL terminator in addrlen
In the utility getinfo path, the address the shm provider
creates in smr_resolve_addr gets duplicated to the info
entry through memdup which does not include the NULL
terminator. The shm provider needs to include a byte for
the NULL terminator and include it in the addrlen so that
memdup duplicates the NULL terminator, avoiding a corrupted
endpoint name later on when creating the shared memory region.

Fix the same issue in the getname path.

Cherry-picked from commit cc93726c9e10979d89b75b033aba34661f20a326

Signed-off-by: aingerson <alexia.ingerson@intel.com>