prov/shm: fix av_insert addr cast
For FI_ADDR_STR, the addr parameter in fi_av_insert
is a list of strings (such as returned by fi_getname), not
a list of smr_addrs. This worked when inserting one address
since the name member is the first element of the struct, but
when inserting more than one address, running through the
list does not work properly. Use the parameter as a list of
strings instead.

Also offset the length by one for the NULL terminator as fixed
in commit cc93726c9e10979d89b75b033aba34661f20a326

Signed-off-by: aingerson <alexia.ingerson@intel.com>