util/bufpool: Ensure alignment is a power of 2
ofi_memalign requires that the alignment parameter be a
power of 2.  However, the buffer pool alignment attribute
may not be, particularly if the user does not specify a
given alignment, or doesn't request a power of 2 alignment.

When allocating new buffer regions, adjust the regions so
that they align on the next power of 2 alignment value.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>