prov/shm: move ep_name_list initialization/cleanup
DL builds do not call ini/fini and will not properly
initialize/cleanup the ep_name_list.

Move the initialization out of ini and use DEFINE_LIST
to initialize.

Make common smr_cleanup function for code that uses
the utility shm code to cleanup the ep_name_list and
call this in the EFA provider which uses the shm provider.

Signed-off-by: aingerson <alexia.ingerson@intel.com>