fabtests/cq_test: limit max CQs to test
A Mellanox bug is causing this test to take too long and
resulting in memory issues/leaks when we try to open too
many CQs (around ~65k). Add a default test max of 32K
to not run into system memory issues but add a command
line argument to override it and test more.

Signed-off-by: aingerson <alexia.ingerson@intel.com>