util/mr_cache: Release lock when flushing cache entries
Flushing cached entries deregisters memory and frees data
buffers.  This can result in uffd events, which will hang
the flushing thread until the uffd thread reads the events.
However, the uffd thread must first acquire the monitor
lock before reading the events.  If the flushing thread
holds the lock, the threads will hang.

Modify the flushing code to release the monitor lock when
performing the actual deregistration.  The lock is only
held to serialize access to the lru list.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>