util/mr_cache: Move flush from create up to caller
In order to redo the locking around flushing entries, we
need to only call flush from the top level API.

The call path to flush is one of the following:

search -> flush
search -> create -> flush
search -> merge -> create -> flush

In all cases, flushing of the cache is only performed
if we are unable to create a new memory region.  We can
handle this failure in search, by having it perform the
flush, then retry the request.  This removes the flush
from the create path.  For the merge path, we simply
pass the failure back to the caller.  This merge path
already flushes additional registrations as part of
its operation, so additional flushing may, but is unlikely,
to result in success.

This change allows us to eventually rework the locking
around the flush operation in a straightforward way.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>