fi_info: Add filter option when displaying environment vars
The number of environment variables can be fairly long.
Using grep on the output can help search for a variable
containing a specific string, but in doing so, it loses
the vars description.  Allow passing in a string to use
as a filter.  Only variables containing that string will
be displayed.

This removes the need for the variable prefix hack to
filter based on provider name.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>