prov/cxi: Implement "standard MRs"
A standard MR is implemented with an LE describing the user buffer
linked to a shared, matching PtlTE. Match bits encode the MR key. The
count of supported standard MRs scales with the number of LEs available,
as opposed to the number of PtlTEs as with optimized MRs.

RMA and AMO interfaces are enhanced to support targetting standard MRs.
Since standard MRs are implemented with a matching PtlTE interface,
there are restrictions to using them. There are no IDC commands to
target a matching interface with a non-zero remote offset. Therefore,
RMA and AMO interfaces use DMA commands when targetting a standard MR.

Inject-style interfaces currently rely on IDC commands. Injects
targetting a standard MR are currently disabled. Software inject support
(using DMA commands) will be added later. The behavior of all other RMA
and AMO transfers should be the same regardless of the type of MR
targetted.

Standard MRs are used implicitly when the requested MR key is greater
than 200. RMA and AMO tests were enhanced to try all operations using a
standard and optimized MR.

Signed-off-by: Zach Tiffany <ztiffany@cray.com>