prov/shm: refactor ptrace_scope to use globally
The availability of CMA/ptrace_scope needs to be
known throughout shm and not just on getinfo since
it is needed to determine whether to use the fast
RMA path or not. Add an environment variable to
access this throughout shm. This also allows
applications to toggle to turn off CMA if desired.

Signed-off-by: aingerson <alexia.ingerson@intel.com>