prov/efa: Set inject buffer size only for rdm
FI_INJECT is not optional, only way to disable it is by setting
tx_attr->inject_size to 0. Setting inject size for dgram endpoint
causes a failure with fi_dgram_pingpong, as there is no way to
discard completions in EFA layer. We should set inject size for rdm
only

Signed-off-by: Dipti Kothari <dkothar@amazon.com>