prov/cxi: Handle Reply event first w/multi-recv
Handle the case where a Reply event is the first to delivered at the
target during a offloaded rendezvous multi-receive. This started
occur frequently after a recent driver fix that caused event combining
to be consistently enabled.

When a Reply event arrives first, set a child request's initiator and
rendezvous ID fields only. These fields are needed for matching to other
events for the transaction. Do not populate other fields of the child
request until the first target event is delivered.

Signed-off-by: Zach Tiffany <ztiffany@cray.com>