prov/efa: Provide explicit warning when rdma-core v27 is not available
The current behavior does not make it obvious why configure failed when
the necessary direct verbs are not available, and needed digging in the
config.log. This commit it makes it explicit for users. The check itself
is fine as it is, given we are checking for the functionality needed and
not for a specific version.

Signed-off-by: Raghu Raja <craghun@amazon.com>