prov/verbs: Refactor MR allocation.
Instead of having two sets of MR operations defined for caching or not
caching MRs, have a single set of MR operations defined and determine if
caching should be used based on a bool cache variable.

By decoupling the setting of the MR operations based on the status of
the cache initialization, new caches can easily be defined for device
memory.

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>