prov/verbs: Fix segfault processing completions
Fixes #5653

If an operation completes in error, then the only valid field
in a verbs work completion is the wr_id.  In order to determine
if a work completion corresponds to a send or receive, we need
to track it ourselves.  Allocate a vrb_context structure for
all posted receive operations (similar to the send side).  This
allows casting the wr_id to a vrb_context always to access the
necessary data stored prior to posting the work request.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>