prov/sockets: Fix fabric and domain name settings
Convert sockets to use the new IP provider getinfo code.  This
ensures that the fabric and domain names are set correctly based
on the addresses used.

Because the common code verifies and formats the fi_info output,
we can remove a bunch of code in the socket provider that performs
similar checks done by the common code.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>