prov/cxi: Use prealloc'd source buffer for SW rdzv
When using a combo of MANAGE_LOCAL and USE_ONCE in an LE, Cassini will
ignore the remote_offset field in a Put and treat offset as 0 (for
USE_ONCE LEs, variable offset for MANAGE_LOCAL LEs). The Put
remote_offset is available in COMM events (not Overflow events). This
allows Software to use the Put remote_offset field for protocol data.

Store the rendezvous Get offset in remote_offset during the SW
rendezvous protocol. Use the offset at the target when performing a Get
to a pre-allocated source buffer. This is similar to how Gets work when
using RPut.

For NETCASSINI-1935.

Signed-off-by: Zach Tiffany <ztiffany@cray.com>