prov/mrail: Properly handle fi_getinfo() with empty hints
The FI_SOURCE bit is added to the hints when getting information from the
core providers. If the original hints is NULL or with empty caps, the
modified hints would result in the returned fi_info structure having only
FI_SOURCE and secondary capability bits set. This is different from the
default behavior with NULL hints or empty caps.

Fix by calling fi_getinfo() twice, with empty caps passed for the first
call to retrieve the default caps.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>