prov/util: Store interface and network names for IP addresses
When obtaining a list of IP addresses available in the system,
also record which interface is associated with the name, along with
the network address.

This data will be needed to fix a regression in the socket provider's
fabric and domain names.  But can also be used to set the fabric and
domain names for the tcp and udp providers.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>