prov/util: Create a common getinfo handler for IP based providers
The tcp and udp providers take the same action when handing getinfo.
Move the duplicated code into the util code.  The same code can be
used by the sockets provider.

This fixes a regression in the sockets provider reporting the correct
domain and fabric names when calling fi_getinfo with NULL hints.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>