prov/util: Map fabric and domain names to src_addr for IP providers
If a src_addr has been specified in the hints to fi_getinfo, we
need to map the address to get the correct network and interface
names.  Update the util getinfo code path for IP based providers
to set the fabric and domain names.

The code currently does this if hints is NULL.  This adds the same
functionality for non-null hints.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>