prov/shm: add FI_DELIVERY_COMPLETE support
If application requests FI_DELIVERY_COMPLETE support,
force the operation to go through the a path that requires
a receive side ACK.

For normal, tagged, and RMA messages,
this means going through the large message CMA path (and
avoiding the fast RMA path as well).

For atomics, this involves going through the atomic fetch path.

Signed-off-by: aingerson <alexia.ingerson@intel.com>