prov/efa: Make rxr_inline_mr_reg() usable beyond the MR cache path
The original rxr_inline_mr_reg() was written with the MR cache usage in
mind and a world where the EFA provider did not require memory
registrations, with the upper layer use pre-registered bounce buffers.
In preparation for FI_MR_HMEM, this commit extends the function to be
more useful beyond the inline registration. It can still walk the iov to
register buffers inline, but will now do it only when a descriptor is
not provided.

Signed-off-by: Raghu Raja <craghun@amazon.com>
Signed-off-by: Wei Zhang <wzam@amazon.com>